refactoring of rendering functions
