+++ src/routes/category-property/[id]/+page.server.ts
@@ -1,5 +1,5 @@
 import { error } from '@sveltejs/kit'
-import { render_nested_formulas } from '$lib/server/rendering'
+import { render_nested_formulas } from '$lib/server/formulas'
 import { decode_property_ID } from '$lib/commons/property.url'
 import { batch } from '$lib/server/db.catdat'
 import sql from 'sql-template-tag'
