+++ src/routes/category/[id]/+page.server.ts
@@ -1,5 +1,5 @@
 import { error } from '@sveltejs/kit'
-import { render_nested_formulas } from '$lib/server/rendering'
+import { render_nested_formulas } from '$lib/server/formulas'
 import { batch } from '$lib/server/db.catdat'
 import sql from 'sql-template-tag'
 import type {
