+++ src/routes/functor-property/[id]/+page.server.ts
@@ -6,7 +6,7 @@ import type {
 	FunctorShort,
 } from '$lib/commons/types'
 import { batch } from '$lib/server/db.catdat'
-import { render_nested_formulas } from '$lib/server/rendering'
+import { render_nested_formulas } from '$lib/server/formulas'
 import { display_functor_implication, display_functor_property } from '$lib/server/utils'
 import { error } from '@sveltejs/kit'
 import sql from 'sql-template-tag'
