+++ src/routes/lemma/[id]/+page.server.ts
@@ -1,6 +1,6 @@
 import type { CategoryShort, Lemma } from '$lib/commons/types'
 import { batch } from '$lib/server/db.catdat'
-import { render_nested_formulas } from '$lib/server/rendering'
+import { render_nested_formulas } from '$lib/server/formulas'
 import { error } from '@sveltejs/kit'
 import sql from 'sql-template-tag'
